# openapi/LinkShortener.json
{
  "components": {
    "schemas": {
      "CreateUrlRequestBody": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/CreateUrlRequestBody.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "type": "string"
          },
          "long_url": {
            "minLength": 1,
            "type": "string"
          },
          "shortcode": {
            "type": "string"
          }
        },
        "required": [
          "long_url"
        ],
        "type": "object"
      },
      "CreateUrlResponseBody": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/CreateUrlResponseBody.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "long_url": {
            "type": "string"
          },
          "short_url": {
            "type": "string"
          },
          "shortcode": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "shortcode",
          "long_url",
          "short_url"
        ],
        "type": "object"
      },
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "location": {
            "description": "Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id'",
            "type": "string"
          },
          "message": {
            "description": "Error message text",
            "type": "string"
          },
          "value": {
            "description": "The value at the given location"
          }
        },
        "type": "object"
      },
      "ErrorModel": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/ErrorModel.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "detail": {
            "description": "A human-readable explanation specific to this occurrence of the problem.",
            "examples": [
              "Property foo is required but is missing."
            ],
            "type": "string"
          },
          "errors": {
            "description": "Optional list of individual error details",
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "instance": {
            "description": "A URI reference that identifies the specific occurrence of the problem.",
            "examples": [
              "https://example.com/error-log/abc123"
            ],
            "format": "uri",
            "type": "string"
          },
          "status": {
            "description": "HTTP status code",
            "examples": [
              400
            ],
            "format": "int64",
            "type": "integer"
          },
          "title": {
            "description": "A short, human-readable summary of the problem type. This value should not change between occurrences of the error.",
            "examples": [
              "Bad Request"
            ],
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "description": "A URI reference to human-readable documentation for the error.",
            "examples": [
              "https://example.com/errors/example"
            ],
            "format": "uri",
            "type": "string"
          }
        },
        "type": "object"
      },
      "ListUrlsResponseBody": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/ListUrlsResponseBody.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/UrlItem"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "page": {
            "format": "int64",
            "type": "integer"
          },
          "page_size": {
            "format": "int64",
            "type": "integer"
          },
          "total": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "type": "object"
      },
      "MeBody": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/MeBody.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "provider_email": {
            "type": "string"
          },
          "user_id": {
            "format": "int64",
            "type": "integer"
          },
          "user_name": {
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "user_name",
          "provider_email"
        ],
        "type": "object"
      },
      "RegisterRequestBody": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/RegisterRequestBody.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "pre_registration_token": {
            "type": "string"
          },
          "user_name": {
            "maxLength": 50,
            "minLength": 3,
            "type": "string"
          }
        },
        "required": [
          "pre_registration_token",
          "user_name"
        ],
        "type": "object"
      },
      "UpdateUrlRequestBody": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/UpdateUrlRequestBody.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "type": "string"
          },
          "last_updated": {
            "format": "date-time",
            "type": "string"
          },
          "long_url": {
            "minLength": 1,
            "type": "string"
          },
          "shortcode": {
            "type": "string"
          }
        },
        "required": [
          "long_url",
          "last_updated"
        ],
        "type": "object"
      },
      "UpdateUrlResponseBody": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/UpdateUrlResponseBody.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "last_updated": {
            "format": "date-time",
            "type": "string"
          },
          "long_url": {
            "type": "string"
          },
          "short_url": {
            "type": "string"
          },
          "shortcode": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "shortcode",
          "long_url",
          "short_url",
          "last_updated"
        ],
        "type": "object"
      },
      "UrlItem": {
        "additionalProperties": false,
        "properties": {
          "expires_at": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "last_updated": {
            "format": "date-time",
            "type": "string"
          },
          "long_url": {
            "type": "string"
          },
          "short_url": {
            "type": "string"
          },
          "shortcode": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "shortcode",
          "long_url",
          "short_url",
          "last_updated"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "description": "App version: dev",
    "title": "Link Shortener API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/auth/logout": {
      "post": {
        "responses": {
          "204": {
            "description": "No Content",
            "headers": {
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Log out the authenticated user by blacklisting the JWT"
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "get-me",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeBody"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Return the authenticated user's profile info"
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "register-user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequestBody"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content",
            "headers": {
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Complete new-user registration after OAuth callback"
      }
    },
    "/r/{shortcode}": {
      "get": {
        "operationId": "redirect-shortcode",
        "parameters": [
          {
            "in": "path",
            "name": "shortcode",
            "required": true,
            "schema": {
              "maxLength": 6,
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Found",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Redirect to the long URL for a given shortcode"
      }
    },
    "/user/account": {
      "delete": {
        "operationId": "delete-account",
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Permanently delete the authenticated user's account and all associated data"
      }
    },
    "/user/urls": {
      "get": {
        "operationId": "list-user-urls",
        "parameters": [
          {
            "explode": false,
            "in": "query",
            "name": "page",
            "schema": {
              "default": 1,
              "format": "int64",
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "explode": false,
            "in": "query",
            "name": "page_size",
            "schema": {
              "format": "int64",
              "maximum": 100,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListUrlsResponseBody"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "List authenticated user's shortened URLs"
      },
      "post": {
        "operationId": "create-shortened-url",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUrlRequestBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUrlResponseBody"
                }
              }
            },
            "description": "Created"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Create a new shortened URL"
      }
    },
    "/user/urls/{id}": {
      "delete": {
        "operationId": "delete-shortened-url",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "explode": false,
            "in": "query",
            "name": "last_updated",
            "required": true,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Delete an existing shortened URL"
      },
      "patch": {
        "operationId": "update-shortened-url",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUrlRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateUrlResponseBody"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Update an existing shortened URL"
      }
    }
  }
}